Update - Enchant
인챈트 로직 수정,
아이템 스탯 표시 로직 수정

+++ Server/Server/Data/Data.Contents.cs
@@ -633,8 +633,8 @@ namespace Server.Data
     {
         public float probability;
         public ItemOptionType option;
-        public float minValue;
-        public float maxValue;
+        public int minValue;
+        public int maxValue;
     }
     [Serializable]
     public class EnchantData
